Remove all references of separate support api and port test to primeapi directory
