+++ .envrc
@@ -253,7 +253,6 @@ export SERVE_DPS=true
 export SERVE_API_INTERNAL=true
 export SERVE_API_GHC=true
 export SERVE_API_PRIME=true
-export SERVE_API_SUPPORT=true
 
 # CDN
 require CLOUD_FRONT_PRIVATE_KEY "See 'chamber read app-devlocal cloud_front_private_key'"
