+++ pkg/cli/hosts.go
@@ -20,8 +20,6 @@ const (
 	HTTPOrdersServerNameFlag string = "http-orders-server-name"
 	// HTTPPrimeServerNameFlag is the HTTP GHC Prime Server Name Flag
 	HTTPPrimeServerNameFlag string = "http-prime-server-name"
-	// HTTPSupportServerNameFlag is the HTTP GHC Prime Server Name Flag
-	HTTPSupportServerNameFlag string = "http-support-server-name"
 	// HTTPMyServerNameLocal is the HTTP My Server Name for Local
 	HTTPMyServerNameLocal string = "milmovelocal"
 	// HTTPOfficeServerNameLocal is the HTTP Office Server Name for Local
@@ -51,7 +49,6 @@ func InitHostFlags(flag *pflag.FlagSet) {
 	flag.String(HTTPAdminServerNameFlag, HTTPAdminServerNameLocal, "Hostname according to environment.")
 	flag.String(HTTPOrdersServerNameFlag, HTTPOrdersServerNameLocal, "Hostname according to environment.")
 	flag.String(HTTPPrimeServerNameFlag, HTTPPrimeServerNameLocal, "Hostname according to environment.")
-	flag.String(HTTPSupportServerNameFlag, HTTPSupportServerNameLocal, "Hostname according to environment")
 }
 
 // CheckHosts validates the Hosts command line flags
@@ -63,7 +60,6 @@ func CheckHosts(v *viper.Viper) error {
 		HTTPAdminServerNameFlag,
 		HTTPOrdersServerNameFlag,
 		HTTPPrimeServerNameFlag,
-		HTTPSupportServerNameFlag,
 	}
 
 	for _, c := range hostVars {
