--- pkg/handlers/supportapi/internal/payloads/model_to_payload_test.go
@@ -1,12 +1,0 @@
-package payloads
-
-import (
-	"testing"
-
-	"github.com/transcom/mymove/pkg/models"
-)
-
-func TestMoveOrder(t *testing.T) {
-	moveOrder := &models.MoveOrder{}
-	MoveOrder(moveOrder)
-}
