--- pkg/handlers/supportapi/move_task_order.go
@@ -1,49 +1,0 @@
-package supportapi
-
-import (
-	"github.com/gofrs/uuid"
-	"go.uber.org/zap"
-
-	"github.com/transcom/mymove/pkg/handlers/supportapi/internal/payloads"
-
-	"github.com/go-openapi/runtime/middleware"
-
-	"github.com/transcom/mymove/pkg/gen/ghcmessages"
-	movetaskorderops "github.com/transcom/mymove/pkg/gen/supportapi/supportoperations/move_task_order"
-	"github.com/transcom/mymove/pkg/handlers"
-	"github.com/transcom/mymove/pkg/services"
-)
-
-// UpdateMoveTaskOrderStatusHandlerFunc updates the status of a Move Task Order
-type UpdateMoveTaskOrderStatusHandlerFunc struct {
-	handlers.HandlerContext
-	moveTaskOrderStatusUpdater services.MoveTaskOrderUpdater
-}
-
-// Handle updates the status of a MoveTaskOrder
-func (h UpdateMoveTaskOrderStatusHandlerFunc) Handle(params movetaskorderops.UpdateMoveTaskOrderStatusParams) middleware.Responder {
-	_, logger := h.SessionAndLoggerFromRequest(params.HTTPRequest)
-	eTag := params.IfMatch
-
-	moveTaskOrderID := uuid.FromStringOrNil(params.MoveTaskOrderID)
-
-	mto, err := h.moveTaskOrderStatusUpdater.MakeAvailableToPrime(moveTaskOrderID, eTag)
-
-	if err != nil {
-		logger.Error("supportapi.MoveTaskOrderHandler error", zap.Error(err))
-		switch err.(type) {
-		case services.NotFoundError:
-			return movetaskorderops.NewUpdateMoveTaskOrderStatusNotFound().WithPayload(&ghcmessages.Error{Message: handlers.FmtString(err.Error())})
-		case services.InvalidInputError:
-			return movetaskorderops.NewUpdateMoveTaskOrderStatusBadRequest().WithPayload(&ghcmessages.Error{Message: handlers.FmtString(err.Error())})
-		case services.PreconditionFailedError:
-			return movetaskorderops.NewUpdateMoveTaskOrderStatusPreconditionFailed().WithPayload(&ghcmessages.Error{Message: handlers.FmtString(err.Error())})
-		default:
-			return movetaskorderops.NewUpdateMoveTaskOrderStatusInternalServerError().WithPayload(&ghcmessages.Error{Message: handlers.FmtString(err.Error())})
-		}
-	}
-
-	moveTaskOrderPayload := payloads.MoveTaskOrder(mto)
-
-	return movetaskorderops.NewUpdateMoveTaskOrderStatusOK().WithPayload(moveTaskOrderPayload)
-}
