--- pkg/handlers/supportapi/move_task_order_test.go
@@ -1,43 +1,0 @@
-package supportapi
-
-import (
-	"net/http/httptest"
-
-	"github.com/transcom/mymove/pkg/etag"
-	"github.com/transcom/mymove/pkg/services/query"
-
-	"github.com/go-openapi/strfmt"
-
-	"github.com/transcom/mymove/pkg/gen/supportapi/supportoperations/move_task_order"
-	movetaskorderops "github.com/transcom/mymove/pkg/gen/supportapi/supportoperations/move_task_order"
-	"github.com/transcom/mymove/pkg/handlers"
-	movetaskorder "github.com/transcom/mymove/pkg/services/move_task_order"
-	"github.com/transcom/mymove/pkg/testdatagen"
-)
-
-func (suite *HandlerSuite) TestUpdateMoveTaskOrderHandlerIntegrationSuccess() {
-	moveTaskOrder := testdatagen.MakeMoveTaskOrder(suite.DB(), testdatagen.Assertions{})
-
-	request := httptest.NewRequest("PATCH", "/move-task-orders/{moveTaskOrderID}/status", nil)
-	params := move_task_order.UpdateMoveTaskOrderStatusParams{
-		HTTPRequest:     request,
-		MoveTaskOrderID: moveTaskOrder.ID.String(),
-		IfMatch:         etag.GenerateEtag(moveTaskOrder.UpdatedAt),
-	}
-	context := handlers.NewHandlerContext(suite.DB(), suite.TestLogger())
-	queryBuilder := query.NewQueryBuilder(suite.DB())
-
-	// make the request
-	handler := UpdateMoveTaskOrderStatusHandlerFunc{context,
-		movetaskorder.NewMoveTaskOrderUpdater(suite.DB(), queryBuilder),
-	}
-	response := handler.Handle(params)
-
-	suite.IsNotErrResponse(response)
-	moveTaskOrdersResponse := response.(*movetaskorderops.UpdateMoveTaskOrderStatusOK)
-	moveTaskOrdersPayload := moveTaskOrdersResponse.Payload
-
-	suite.Assertions.IsType(&move_task_order.UpdateMoveTaskOrderStatusOK{}, response)
-	suite.Equal(moveTaskOrdersPayload.ID, strfmt.UUID(moveTaskOrder.ID.String()))
-	suite.Equal(*moveTaskOrdersPayload.IsAvailableToPrime, true)
-}
