+++ scripts/gen-server
@@ -71,11 +71,9 @@ generate_server swagger/ghc.yaml      ghcmessages      ghcapi      ghcoperations
 generate_server swagger/internal.yaml internalmessages internalapi internaloperations
 generate_server swagger/orders.yaml   ordersmessages   ordersapi   ordersoperations
 generate_server swagger/prime.yaml    primemessages    primeapi    primeoperations
-generate_server swagger/support.yaml supportmessages supportapi supportoperations
 
 # Generate Client Code
 generate_client swagger/prime.yaml primemessages primeclient
-generate_client swagger/support.yaml supportmessages supportclient
 
 # Wait for all processes to finish
 for pid in ${pids[*]}; do
