+++ src/setupProxy.js
@@ -6,7 +6,6 @@ module.exports = function (app) {
   app.use(createProxyMiddleware('/admin', { target: 'http://milmovelocal:8080/' }));
   app.use(createProxyMiddleware('/ghc', { target: 'http://milmovelocal:8080/' }));
   app.use(createProxyMiddleware('/prime', { target: 'http://milmovelocal:8080/' }));
-  app.use(createProxyMiddleware('/support', { target: 'http://milmovelocal:8080/' }));
   app.use(createProxyMiddleware('/storage', { target: 'http://milmovelocal:8080/' }));
   app.use(createProxyMiddleware('/devlocal-auth', { target: 'http://milmovelocal:8080/' }));
   app.use(createProxyMiddleware('/auth/**', { target: 'http://milmovelocal:8080/' }));
